+++ custom_components/wattio/switch.py
@@ -6,11 +6,15 @@ try:
 except ImportError:
     from homeassistant.components.switch import SwitchDevice as SwitchEntity
 
+from homeassistant.helpers.aiohttp_client import async_get_clientsession
+
 from homeassistant.const import STATE_ON
 
 from . import WattioDevice, wattioApi
 
 from .const import CONF_EXCLUSIONS, DOMAIN, ICON, SECURITY, SWITCHES
+
+
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -70,23 +74,23 @@ class WattioSwitch(WattioDevice, SwitchEntity):
 
     async def async_turn_on(self):
         """Turn On method."""
-        wattio = wattioApi(self.hass.data[DOMAIN]["token"])
+        wattio = wattioApi(self.hass.data[DOMAIN]["token"], async_get_clientsession(self.hass))
         # Cambiar a debug
         _LOGGER.debug(
             "Sending ON request to SWITCH device %s (%s)", self._ieee, self._name
         )
-        wattio.set_switch_status(self._ieee, "on", self._devtype)
+        await wattio.async_set_switch_status(self._ieee, "on", self._devtype)
         self._state = True
         self.schedule_update_ha_state()
 
     async def async_turn_off(self):
         """Turn Off method."""
-        wattio = wattioApi(self.hass.data[DOMAIN]["token"])
+        wattio = wattioApi(self.hass.data[DOMAIN]["token"], async_get_clientsession(self.hass))
         # Cambiar a debug
         _LOGGER.debug(
             "Sending OFF request to SWITCH device %s (%s)", self._ieee, self._name
         )
-        wattio.set_switch_status(self._ieee, "off", self._devtype)
+        await wattio.async_set_switch_status(self._ieee, "off", self._devtype)
         self._state = False
         self.schedule_update_ha_state()
 
@@ -175,21 +179,21 @@ class WattioSecurity(WattioDevice, SwitchEntity):
 
     async def async_turn_on(self):
         """Turn On method."""
-        wattio = wattioApi(self.hass.data[DOMAIN]["token"])
+        wattio = wattioApi(self.hass.data[DOMAIN]["token"], async_get_clientsession(self.hass))
         _LOGGER.debug(
             "Sending ON request to SECURITY device %s (%s)", self._ieee, self._name
         )
-        wattio.set_security_device_status(self._devtype, self._ieee, "on")
+        await wattio.async_set_security_device_status(self._devtype, self._ieee, "on")
         self._state = True
         self.schedule_update_ha_state()
 
     async def async_turn_off(self):
         """Turn Off method."""
-        wattio = wattioApi(self.hass.data[DOMAIN]["token"])
+        wattio = wattioApi(self.hass.data[DOMAIN]["token"], async_get_clientsession(self.hass))
         _LOGGER.debug(
             "Sending OFF request to SECURITY device %s (%s)", self._ieee, self._name
         )
-        wattio.set_security_device_status(self._devtype, self._ieee, "off")
+        await wattio.async_set_security_device_status(self._devtype, self._ieee, "off")
         self._state = False
         self.schedule_update_ha_state()